feat(placeholders): Implement conditional placeholders and their matching logic

+++ src/main/java/com/zerog/neoessentials/ui/tab/TabManager.java
@@ -146,13 +146,12 @@ public class TabManager {
         initialized = true;
         NeoEssentials.LOGGER.info("TabManager initialized");
     }
-    
-    /**
+      /**
      * Loads configuration settings for the tablist
      */
     public void loadConfig() {
         // Load general settings
-        updateInterval = TablistTomlConfig.UPDATE_INTERVAL.get();
+        updateInterval = TablistTomlConfig.UPDATE_INTERVAL.get().intValue(); // Convert Long to int
         
         // Each feature loads its own config
         headerFooterFeature.loadConfig();
